Legend looks better with themes manually reordered to match rp2040pins.csv
Idea here is that the legend boxes match the pin mux boxes from closest-to-board to furthest. I tried looking at doing this automatically, but because the themes and MUX names don’t match (they’re remapped in the box-drawing loop) it’s just not that simple. So this looks OK for RP2040 boards but might not work as well for others (though still functional).